truncation selection + host growth

+++ source/sgp_mode/SGPSymbiont.h
@@ -55,7 +55,7 @@ public:
     }
     // Invalidate any in-progress reproduction
     if (cpu.state.in_progress_repro != -1) {
-      if(my_world->to_reproduce.size() < cpu.state.in_progress_repro) {
+      if(my_world->to_reproduce.size() < static_cast<size_t>(cpu.state.in_progress_repro)) {
         my_world->to_reproduce[cpu.state.in_progress_repro].second =
           emp::WorldPosition::invalid_id;
       }
